partial support for new GitHub review flow
